Pass original proxy objects from Subscribe to callbacks
Instead of creating a new proxy object using the internal proxy from the
thin bindings layer, store the thick proxy object provided to function
Subscribe in the user data when adding a listener. This way we can
retrieve it again when event occurs.

The way for a user to attach additional data to a proxy object is by
creating a new tagged type inheriting the proxy type.

For example:

   type My_Registry is new Wayland.Client.Protocol.Registry with record
      Additional_Data : Data_Type;
   end record;