Use errno from C99 standard to determine whether to poll (again)
Function Display_Flush may return -1 and set errno to EAGAIN if not all
data could be flushed. The Wayland documentation says in that case the
file descriptor must be polled and then try calling Display_Flush again.

According to the man page of poll(2), the function may return EINTR or
EAGAIN, and in this case poll() must be called again in a loop.

Signed-off-by: onox <denkpadje@gmail.com>