Add procedure Move for type Data_Offer in thick bindings
For a Data_Offer object to be usable it needs to be saved so that it can
be used later when the user presses Ctrl + v to initiate pasting
something from the clipboard. Because the type Data_Offer is limited,
the internal Proxy must be moved to a new object.

Signed-off-by: onox <denkpadje@gmail.com>